backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
passlib[bcrypt]
bcrypt
python-jose[cryptography]
requests
groq
pydantic
email-validator
gunicorn
